Add files via upload
This is my first project: 
In this i made an number game in which user can play multiple times as much user wants.
In this game user have to guess a random number that is between 1 and 100 if the user guess the number correctly. The user will won the game. But there is a catch user will only have 10 attempts to guess the number after using all attemps user will lose the game.